+++ kensho_kenverters/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Remove table validation for rows and columns to not fail downstream of Extract model failures
 
+Added support of figure extracted table.
+
 ## v1.2.2
 
 Use underscores instead of hyphens for setuptools 78