+++ kensho_kenverters/CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changelog
 
+## v1.2.4
+
+Add conversion from a given table annotation to grid, finding the first and last associated text object
+
 ## v1.2.3
 
 * Remove table validation for rows and columns to not fail downstream of Extract model failures
-
 * Added support of figure extracted table.
 
 ## v1.2.2